// archived/Trading.Backtest.Web/wwwroot/src/types/api.ts
export interface Trade {
  id: string;
  direction: string;
  openTime: string;
  openPrice: number;
  stopLoss: number;
  takeProfit: number;
  stopLossPips: number;
  takeProfitPips: number;
  closeTime?: string;
  closePrice?: number;
  closeReason?: string;
  profitLoss: number;
  returnRate: number;
  lots: number;
}